--- a/backend/app/build.gradle.kts
+++ b/backend/app/build.gradle.kts
@@ -1,0 +1,35 @@
+plugins {
+    kotlin("jvm")
+    kotlin("plugin.spring")
+    kotlin("plugin.jpa")
+    kotlin("kapt")
+    id("org.springframework.boot")
+}
+
+dependencies {
+    implementation(project(":data"))
+    implementation(project(":security"))
+    implementation(project(":api"))
+    
+    implementation("org.springframework.boot:spring-boot-starter-web")
+    implementation("org.springframework.boot:spring-boot-starter-validation")
+    implementation("org.springframework.boot:spring-boot-starter-security")
+    implementation("org.springframework.boot:spring-boot-starter-actuator")
+    
+    implementation("com.fasterxml.jackson.module:jackson-module-kotlin")
+    implementation("org.jetbrains.kotlin:kotlin-reflect")
+    
+    testImplementation("org.springframework.boot:spring-boot-starter-test") {
+        exclude(group = "org.junit.vintage", module = "junit-vintage-engine")
+    }
+    testImplementation("org.springframework.security:spring-security-test")
+}
+
+tasks.getByName<org.springframework.boot.gradle.tasks.bundling.BootJar>("bootJar") {
+    enabled = true
+    mainClass.set("io.tolgee.TolgeeApplicationKt")
+}
+
+tasks.getByName<Jar>("jar") {
+    enabled = false
+}