… .github/workflows/test-optimization.yml
… backend/.java-version
… backend/README.md
… backend/application.yml
… backend/data/src/main/kotlin/io/tolgee/model/Project.kt
… backend/data/src/main/kotlin/io/tolgee/model/contentDelivery/AzureContentStorageConfig.kt
… backend/data/src/main/kotlin/io/tolgee/model/contentDelivery/ContentDeliveryConfig.kt
… backend/data/src/main/kotlin/io/tolgee/model/contentDelivery/ContentStorage.kt
… backend/data/src/main/kotlin/io/tolgee/model/contentDelivery/S3ContentStorageConfig.kt
… backend/data/src/main/kotlin/io/tolgee/repository/contentDelivery/ContentDeliveryConfigRepository.kt
… backend/data/src/main/kotlin/io/tolgee/repository/contentDelivery/OptimizedContentDeliveryConfigRepository.kt
… backend/data/src/main/kotlin/io/tolgee/service/contentDelivery/ContentDeliveryConfigService.kt
… backend/data/src/test/kotlin/io/tolgee/AbstractDatabaseTest.kt
… backend/data/src/test/kotlin/io/tolgee/TestDataFactory.kt
… backend/data/src/test/kotlin/io/tolgee/TestSuites.kt
… backend/data/src/test/kotlin/io/tolgee/config/FastTestContextInitializer.kt
… backend/data/src/test/kotlin/io/tolgee/config/HibernateOptimizationConfig.kt
… backend/data/src/test/kotlin/io/tolgee/config/TestApplicationRunner.kt
… backend/data/src/test/kotlin/io/tolgee/config/TestDatabaseInitializer.kt
… backend/data/src/test/kotlin/io/tolgee/config/TestExecutionConfig.kt
… backend/data/src/test/kotlin/io/tolgee/config/TestSliceConfiguration.kt
… backend/data/src/test/kotlin/io/tolgee/configuration/HibernateStatisticsConfiguration.kt
… backend/data/src/test/kotlin/io/tolgee/configuration/LazyInitConfiguration.kt
… backend/data/src/test/kotlin/io/tolgee/configuration/LightweightIntegrationTest.kt
… backend/data/src/test/kotlin/io/tolgee/configuration/LightweightTestConfig.kt
… backend/data/src/test/kotlin/io/tolgee/configuration/OptimizedTestConfiguration.kt
… backend/data/src/test/kotlin/io/tolgee/configuration/TestApplicationConfig.kt
… backend/data/src/test/kotlin/io/tolgee/configuration/TestBeanPostProcessor.kt
… backend/data/src/test/kotlin/io/tolgee/configuration/TestContextConfiguration.kt
… backend/data/src/test/kotlin/io/tolgee/configuration/TestDatabaseConfig.kt